Split up dashboard spec, and other desc changes.
