+++ spec/features/user_authentication_feature_spec.rb
@@ -1,11 +1,11 @@
 require 'rails_helper'
 
-RSpec.feature 'user authentication' do
-  context  'cso with authentication' do
+RSpec.feature 'User authentication' do
+  context 'a Custody Suite Officer' do
     let(:cso_password) { 'password' }
     let(:cso_user) { User.create(email: 'cso@example.com', password: cso_password, role: :cso) }
 
-    scenario 'logs in with valid credentials' do
+    specify 'can log in with valid credentials' do
       visit new_user_session_path
       fill_in 'user_email', with: cso_user.email
       fill_in 'user_password', with: cso_password
@@ -13,7 +13,7 @@ RSpec.feature 'user authentication' do
       expect(page).to have_content('Custody Suite Officer Dashboard')
     end
 
-    scenario 'logs in with invalid credentials' do
+    specify 'cannot log in with invalid credentials' do
       visit new_user_session_path
       fill_in 'user_email', with: cso_user.email
       fill_in 'user_password', with: 'notarealpassword'
@@ -22,11 +22,11 @@ RSpec.feature 'user authentication' do
     end
   end
 
-  context 'cco with authentication' do
+  context 'a Call Center Operative' do
     let(:cco_password) { 'password' }
     let(:cco_user) { User.create(email: 'cco@example.com', password: cco_password, role: :cco) }
 
-    scenario 'logs in with valid credentials' do
+    specify 'can log in with valid credentials' do
       visit new_user_session_path
       fill_in 'user_email', with: cco_user.email
       fill_in 'user_password', with: cco_password
@@ -34,7 +34,7 @@ RSpec.feature 'user authentication' do
       expect(page).to have_content('Call Center Operative Dashboard')
     end
 
-    scenario 'logs in with invalid credentials' do
+    specify 'cannot log in with invalid credentials' do
       visit new_user_session_path
       fill_in 'user_email', with: cco_user.email
       fill_in 'user_password', with: 'notarealpassword'
@@ -43,11 +43,11 @@ RSpec.feature 'user authentication' do
     end
   end
 
-  context 'solicitor with authentication' do
+  context 'a Solicitor' do
     let(:solicitor_password) { 'password' }
     let(:solicitor_user) { User.create(email: 'solicitor@example.com', password: solicitor_password, role: :solicitor) }
 
-    scenario 'logs in with valid credentials' do
+    specify 'can log in with valid credentials' do
       visit new_user_session_path
       fill_in 'user_email', with: solicitor_user.email
       fill_in 'user_password', with: solicitor_password
@@ -55,20 +55,12 @@ RSpec.feature 'user authentication' do
       expect(page).to have_content('Solicitor Dashboard')
     end
 
-    scenario 'logs in with invalid credentials' do
+    specify 'cannot log in with invalid credentials' do
       visit new_user_session_path
       fill_in 'user_email', with: solicitor_user.email
       fill_in 'user_password', with: 'notarealpassword'
       click_button 'Sign in'
       expect(page).to_not have_content('Solcitior Dashboard')
-    end
-  end
-
-
-  context 'unauthenticated' do
-    scenario 'logs in with valid credentials' do
-      visit root_path
-      expect(page).to have_button('Sign in')
     end
   end
 end
